adcionando sala com html atualizado etc

+++ src/app/components/sala/sala.component.ts
@@ -8,7 +8,6 @@ import { SalaEquipamento, SalaModel, TipoSala } from './../../shared/model/sala.
 import { Component, OnInit } from '@angular/core';
 import { MenuItem, MessageService, SelectItem, ConfirmationService } from 'primeng/api';
 
-
 @Component({
   selector: 'app-sala',
   templateUrl: './sala.component.html',
@@ -40,7 +39,7 @@ export class SalaComponent implements OnInit {
     equipamentos: null
   };
 
-  
+
 
   ngOnInit(): void {
     this.getAll();
@@ -61,16 +60,25 @@ export class SalaComponent implements OnInit {
     this.tiposSalas =
       [
         { label: 'Tipo Sala: ', value: null },
-        { label: 'sala de reunião', value: 1 },
-        { label: 'sala de trabalho', value: 2 },
-        { label: 'sala de video', value: 3 },
-        { label: 'sala de palestras', value: 4 },
-        { label: 'auditorio', value: 5 },
+        { label: 'Sala de reunião', value: 1 },
+        { label: 'Sala de trabalho', value: 2 },
+        { label: 'Sala de video', value: 3 },
+        { label: 'Sala de palestras', value: 4 },
+        { label: 'Auditório', value: 5 },
       ]
   }
 
   confirm(sala: SalaModel) {
-    this.deleteWithButton(sala);
+    this.confirmationService.confirm({
+      message: 'Você tem certeza que deseja excluir esta sala?',
+      accept: () => {
+        this.deleteWithButton(sala);
+      },
+      reject: () => {
+        this.displayEditDialog = false;
+        this.displaySaveDialog = false;
+      }
+    });
   }
   deleteWithButton(sala: SalaModel) {
     this.salaService.delete(sala).subscribe(
@@ -104,11 +112,6 @@ export class SalaComponent implements OnInit {
     });
   }
 
-  getSalaEquipamentoDialog()
-  {
-    
-  }
-
   getAll() {
     this.salaService.getSalas().subscribe(
       (result: any) => {
@@ -135,6 +138,27 @@ export class SalaComponent implements OnInit {
     )
   }
 
+  editEquipWithButton() {
+    this.sala.equipamentos = this.salaEquipamentoService.getSalaEquipamentos();
+    this.salaService.edit(this.sala).subscribe(
+      () => {
+        this.displayEditDialog = false;
+        this.sala = new SalaModel;
+        this.messageService.add({
+          severity: 'success',
+          summary: "Resultado", detail: "Equipamento editado com sucesso"
+        });
+        this.getAll();
+      },
+      error => {
+        this.messageService.add({
+          severity: 'error', summary: "Error",
+          detail: "Equipamento não pode ser editado, verifique os dados e tente novamente"
+        })
+      }
+    )
+  }
+
   getTipoSala(id: number) {
     return TipoSala[id];
   }
@@ -149,5 +173,8 @@ export class SalaComponent implements OnInit {
   }
 
   showEditDialog(sala: SalaModel) {
+    this.displayEditDialog = true;
+    this.sala = sala;
+    this.getAll();
   }
 }
